+++ packages/vinext/src/cloudflare/index.ts
@@ -10,6 +10,6 @@
  * registration. An explicit `setCdnCacheAdapter(...)` always takes precedence.
  */
 
-export { KVCacheHandler } from "./kv-cache-handler.js";
-export { CloudflareCdnCacheAdapter } from "./cloudflare-cdn-cache.js";
+export { KVCacheHandler } from "./cache/kv-data-adapter.runtime.js";
+export { CloudflareCdnCacheAdapter } from "./cache/cdn-adapter.runtime.js";
 export { runTPR, type TPROptions, type TPRResult } from "./tpr.js";
